Add support for timeouts
Reqwest has built in support for timeouts, but the async libraries
default to not having any timeout. When there is a network issue of
some kind, surfaced more when using the library on a cron basis, it is
common to see the processes hanging with the async reqwest clients
never timing out. This sets a default timeout of 10s, while allowing
you to pass in a value as well to override this if you desire a
shorter or longer one. Also included are four unit tests, two to test
that both in the futures and normal calls the default timeout is set
to the default value, and two to test that in both the futures and the
normal calls that when an override is specified it is respected.