Add support for TradeVolume private endpoint
This is an authenticated endpoint, taking one or more pairs comma
separated, eg. "ADAUSD" or "ADAUSD,KAVAUSD" in addition to the
optional fee-info. The response tells the level of volume done in the
last 30 days revolving in addition to current fee and next fee ( if
any ) and next volume level for discount ( if any ).